refactor: inventory(itemstackrequest) (#1481)
* refactor: inventory

* feat: impl TransferItemActionProcessor

* feat: tmp commit

* fix: recipes

* fix: recipes 2

* fix: recipes 3

* fix: recipes 4

* fix: recipes 5

* fix: 6

* fix: 7

* fix: 8

* fix: 9

* fix: 10

* fix: enable action

* feat: use bfs algorithm for shrink ShapedRecipe and match

* fix: typo

* fix: recipe craft

+++ src/main/java/cn/nukkit/network/protocol/ContainerClosePacket.java
@@ -19,7 +19,7 @@ public class ContainerClosePacket extends DataPacket {
 
     @Override
     public void decode() {
-        this.windowId = (byte) this.getByte();
+        this.windowId = this.getByte();
         this.wasServerInitiated = this.getBoolean();
     }
